requirements.txt
django
djangorestframework
django-cors-headers
drf-spectacular
requests
python-dotenv
gunicorn
pytest
pytest-django
